+++ src/components/format-amount/format-amount.tsx
@@ -1,22 +1,28 @@
 import { Component, Prop, h } from '@stencil/core';
+import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 
 @Component({
   tag: 'format-amount',
   styleUrl: 'format-amount.css',
-  shadow: true
+  shadow: true,
 })
 export class FormatAmount {
   @Prop() class?: string;
+  @Prop() dataTestId?: string = DataTestIdsEnum.formatAmountComponent;
   @Prop() isValid: boolean;
   @Prop() label?: string;
   @Prop() labelClass?: string;
   @Prop() valueDecimal: string;
   @Prop() valueInteger: string;
 
+  render() {
+    return this.isValid ? this.renderValid() : this.renderInvalid();
+  }
+
   private renderInvalid() {
     return (
-      <span data-testid='formatAmountComponent' class={this.class}>
-        <span class='int-amount' data-testid='formatAmountInt'>
+      <span data-testid={this.dataTestId} class={this.class}>
+        <span class="int-amount" data-testid={DataTestIdsEnum.formatAmountInt}>
           ...
         </span>
       </span>
@@ -25,31 +31,27 @@ export class FormatAmount {
 
   private renderValid() {
     return (
-      <span data-testid='formatAmountComponent' class={this.class}>
-        <span class='int-amount' data-testid='formatAmountInt'>
+      <span data-testid={this.dataTestId} class={this.class}>
+        <span class="int-amount" data-testid={DataTestIdsEnum.formatAmountInt}>
           {this.valueInteger}
         </span>
         {this.valueDecimal && (
-          <span class='decimals' data-testid='formatAmountDecimals'>
+          <span class="decimals" data-testid={DataTestIdsEnum.formatAmountDecimals}>
             .{this.valueDecimal}
           </span>
         )}
         {this.label && (
           <span
             class={{
-              'symbol': true,
-              [this.labelClass]: Boolean(this.labelClass)
+              symbol: true,
+              [this.labelClass]: Boolean(this.labelClass),
             }}
-            data-testid='formatAmountSymbol'
+            data-testid={DataTestIdsEnum.formatAmountSymbol}
           >
             {this.label}
           </span>
         )}
       </span>
     );
-  }
-
-  render() {
-    return this.isValid ? this.renderValid() : this.renderInvalid();
   }
 }