+++ src/components/transaction-toast/components/transaction-toast-list/transaction-toast-list.types.ts
@@ -1,4 +1,0 @@
-export enum TransactionToastEventsEnum {
-  'CLOSE_TOAST' = 'CLOSE_TOAST',
-  'TRANSACTION_TOAST_DATA_UPDATE' = 'TRANSACTION_TOAST_DATA_UPDATE',
-}